+++ plugin/src/main/kotlin/de/infolektuell/gradle/jextract/extensions/JextractExtension.kt
@@ -3,7 +3,9 @@ package de.infolektuell.gradle.jextract.extensions
 import org.gradle.api.Action
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.model.ObjectFactory
+import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Nested
+import org.gradle.api.tasks.SourceSet
 import javax.inject.Inject
 
 abstract class JextractExtension @Inject constructor(objects: ObjectFactory) {
@@ -20,6 +22,8 @@ abstract class JextractExtension @Inject constructor(objects: ObjectFactory) {
     fun libraries(action: Action<in NamedDomainObjectContainer<LibraryHandler>>) {
         action.execute(libraries)
     }
+    /** The generated Java sources will be added to this source set (main by default) */
+    abstract val sourceSet: Property<SourceSet>
     companion object {
         const val EXTENSION_NAME = "jextract"
     }