+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Create changelog file for release notes
+- The source set the generated sources are added to is now user-selectable via extension property, `main` is chosen by default.
 
 ## [0.2.1] - 2024-09-28